port: fix demo respawn loop - s16 anim sentinels + logical air-steer masks
Root causes found by diffing a PS1 reference trace of the SCIE demo
(make trace STATE=<menu savestate> POKE=0x8009DD88=2,0x8009DD92=1 --
new RAMDUMP_POKE knob pokes direct_level_load/demo_return_flag so the
game itself runs the SetupAndStartLevel demo branch):

1. ApplyPendingSpriteState compared the pending frame/loop/target -1
   sentinels as s32, but the asm uses lh (s16): a pending 0xFFFF never
   matched, so targetFrame stayed -1 and the spawn portal's eject anim
   looped forever, re-firing its embedded 0x100F warp trigger every ~28
   frames -- yanking the player back to the portal (the 'respawn loop').
   With the fix the anim completes, event 2 fires, and the portal
   removes itself exactly like the PS1.
2. PlayerCallback_JumpInputAndCounters air steering tested raw pad bits
   (0x20/0x80) instead of the logical direction masks (0x2000/0x8000,
   asm andi) -- the demo's held run button (0x80) read as LEFT, drifting
   the player off the spawn.

After both fixes the port's demo trajectory matches the PS1 reference
frame-for-frame through spawn, teleporter eject, fall, landing and the
first walk (verified via gdb watchpoints + aligned GameState trace diff
at offset 6771).

Also converted down the chain: PlayerStateInit_BounceActive,
FreeAndCoalesceVRAMSlot (portal removal path),
PlayerCallback_WalkingCollisionHandler, AddToXPositionList; boot logs
the loaded level; PORT_TRACE_PLAYER=1 per-frame player dump.
Next stub: CollectibleOrbTickCallback.

+++ port/decomp/anim/ApplyPendingSpriteState.c
@@ -53,7 +53,9 @@ void ApplyPendingSpriteState(void *entityPtr) {
     flags = *(u16 *)(e + 0xE0);
     if ((flags & ANIM_CHG_FRAME) != 0) {
         if ((flags & 0x200) == 0) {
-            if (*(s32 *)(e + 0xC0) == -1) {
+            /* -1 sentinel test is a SIGNED HALFWORD compare in the asm (lh),
+             * so a pending value written as u16 0xFFFF must match too. */
+            if (*(s16 *)(e + 0xC0) == -1) {
                 *(s16 *)(e + 0xDA) = *(s16 *)(e + 0xD8) - 1;
             } else {
                 *(s16 *)(e + 0xDA) = (s16)*(s32 *)(e + 0xC0);
@@ -66,7 +68,7 @@ void ApplyPendingSpriteState(void *entityPtr) {
 
     if ((flags & ANIM_CHG_LOOP_FRAME) != 0) {
         if ((flags & 0x400) == 0) {
-            if (*(s32 *)(e + 0xC4) == -1) {
+            if (*(s16 *)(e + 0xC4) == -1) {          /* lh in asm, see above */
                 *(s16 *)(e + 0xDC) = *(s16 *)(e + 0xD8) - 1;
             } else {
                 *(s16 *)(e + 0xDC) = (s16)*(s32 *)(e + 0xC4);
@@ -78,7 +80,7 @@ void ApplyPendingSpriteState(void *entityPtr) {
 
     if ((*(u16 *)(e + 0xE0) & ANIM_CHG_TARGET_FRAME) != 0) {
         if ((*(u16 *)(e + 0xE0) & 0x800) == 0) {
-            if (*(s32 *)(e + 0xC8) == -1) {
+            if (*(s16 *)(e + 0xC8) == -1) {          /* lh in asm, see above */
                 *(s16 *)(e + 0xDE) = *(s16 *)(e + 0xD8) - 1;
             } else {
                 *(s16 *)(e + 0xDE) = (s16)*(s32 *)(e + 0xC8);
